root/apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy[asyncio]
asyncpg
faststream[rabbit]
redis
curl_cffi
httpx
qdrant-client